[#22] ColorsWord Page 1차 수정 작업
